Refactor game config loading to service
Introduced IGameConfigService and GameConfigService to encapsulate game configuration loading logic. Updated MainWindowViewModel to use the new service, removing direct JSON handling and related static methods from the view model for better separation of concerns.
