+++ KobberLan/Services/IGameConfigService.cs
@@ -1,0 +1,9 @@
+using KobberLan.Models;
+
+namespace KobberLan.Services;
+
+public interface IGameConfigService
+{
+    public KobberLanConfig? TryLoad(string gameFolder);
+    public string GetGameFullPath();
+}
